+++ module-service/src/main/java/com/checkping/service/project/TaskBoardFileServiceImpl.java
@@ -4,8 +4,8 @@ import com.checkping.common.utils.FileResponse;
 import com.checkping.domain.question.TaskBoard;
 import com.checkping.domain.question.TaskBoardFile;
 import com.checkping.exception.project.TaskBoardNotFoundEntityException;
-import com.checkping.infra.repository.project.taskboard.TaskBoardReader;
-import com.checkping.infra.repository.project.taskboardfile.TaskBoardFileStore;
+import com.checkping.infra.repository.question.TaskBoardReader;
+import com.checkping.infra.repository.question.file.TaskBoardFileStore;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
