+++ module-service/src/main/java/com/checkping/service/project/TaskBoardLinkServiceImpl.java
@@ -5,8 +5,8 @@ import com.checkping.domain.question.TaskBoardLink;
 import com.checkping.dto.TaskBoardLinkRequest;
 import com.checkping.dto.TaskBoardLinkResponse.TaskBoardLinkDto;
 import com.checkping.exception.project.TaskBoardNotFoundEntityException;
-import com.checkping.infra.repository.project.taskboard.TaskBoardReader;
-import com.checkping.infra.repository.project.taskboardlink.TaskBoardLinkStore;
+import com.checkping.infra.repository.question.TaskBoardReader;
+import com.checkping.infra.repository.question.link.TaskBoardLinkStore;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
