+++ module-service/src/main/java/com/checkping/service/project/TaskBoardServiceImpl.java
@@ -13,12 +13,12 @@ import com.checkping.dto.TaskBoardResponse;
 import com.checkping.dto.TaskBoardResponse.TaskBoardItemDto;
 import com.checkping.dto.TaskBoardResponse.TaskBoardListDto;
 import com.checkping.exception.project.TaskBoardNotFoundEntityException;
-import com.checkping.infra.repository.project.taskboard.TaskBoardReader;
-import com.checkping.infra.repository.project.taskboard.TaskBoardStore;
-import com.checkping.infra.repository.project.taskboardcomment.TaskBoardCommentReader;
-import com.checkping.infra.repository.project.taskboardcomment.TaskBoardCommentStore;
-import com.checkping.infra.repository.project.taskboardfile.TaskBoardFileStore;
-import com.checkping.infra.repository.project.taskboardlink.TaskBoardLinkStore;
+import com.checkping.infra.repository.question.TaskBoardReader;
+import com.checkping.infra.repository.question.TaskBoardStore;
+import com.checkping.infra.repository.question.comment.TaskBoardCommentReader;
+import com.checkping.infra.repository.question.comment.TaskBoardCommentStore;
+import com.checkping.infra.repository.question.file.TaskBoardFileStore;
+import com.checkping.infra.repository.question.link.TaskBoardLinkStore;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
